+++ src/filter_feedback_matrix.cpp
@@ -6,7 +6,7 @@
 
 namespace sfFDN
 {
-FilterFeedbackMatrix::FilterFeedbackMatrix(size_t N)
+FilterFeedbackMatrix::FilterFeedbackMatrix(uint32_t N)
     : FeedbackMatrix(N)
 {
     stages_.clear();
@@ -21,38 +21,13 @@ void FilterFeedbackMatrix::Clear()
     }
 }
 
-// void FilterFeedbackMatrix::SetDelays(std::span<size_t> delays)
-// {
-//     assert(delays.size() == N_ * (K_ - 1));
-//     assert(stages_.size() == K_ - 1);
-
-//     for (size_t i = 0; i < stages_.size(); ++i)
-//     {
-//         auto stage_delays = delays.subspan(i * N_, N_);
-//         stages_[i].SetDelays(stage_delays);
-//     }
-// }
-
-// void FilterFeedbackMatrix::SetMatrices(std::span<ScalarFeedbackMatrix> mixing_matrices)
-// {
-//     assert(mixing_matrices.size() == K_);
-//     assert(stages_.size() == K_ - 1);
-
-//     for (size_t i = 0; i < stages_.size(); ++i)
-//     {
-//         stages_[i].SetMatrix(mixing_matrices[i]);
-//     }
-
-//     last_mat_ = mixing_matrices[K_ - 1];
-// }
-
-void FilterFeedbackMatrix::ConstructMatrix(std::span<size_t> delays, std::span<ScalarFeedbackMatrix> mixing_matrices)
+void FilterFeedbackMatrix::ConstructMatrix(std::span<uint32_t> delays, std::span<ScalarFeedbackMatrix> mixing_matrices)
 {
-    const size_t num_stages = mixing_matrices.size();
+    const uint32_t num_stages = mixing_matrices.size();
     assert(delays.size() == N_ * (num_stages - 1));
 
     stages_.reserve(num_stages - 1);
-    for (size_t i = 0; i < num_stages - 1; ++i)
+    for (uint32_t i = 0; i < num_stages - 1; ++i)
     {
         auto stage_delays = delays.subspan(i * N_, N_);
         stages_.emplace_back(N_, stage_delays);
@@ -73,7 +48,7 @@ void FilterFeedbackMatrix::Process(const AudioBuffer& input, AudioBuffer& output
         // Apply first stage
         stages_[0].Process(input, output);
 
-        for (size_t i = 1; i < stages_.size(); ++i)
+        for (uint32_t i = 1; i < stages_.size(); ++i)
         {
             stages_[i].Process(output, output);
         }
