+++ src/filter_feedback_matrix.h
@@ -13,22 +13,22 @@ namespace sfFDN
 class FilterFeedbackMatrix : public FeedbackMatrix
 {
   public:
-    FilterFeedbackMatrix(size_t N);
+    FilterFeedbackMatrix(uint32_t N);
 
     void Clear();
-    // void SetDelays(std::span<size_t> delays);
+    // void SetDelays(std::span<uint32_t> delays);
     // void SetMatrices(std::span<ScalarFeedbackMatrix> mixing_matrices);
 
-    void ConstructMatrix(std::span<size_t> delays, std::span<ScalarFeedbackMatrix> mixing_matrices);
+    void ConstructMatrix(std::span<uint32_t> delays, std::span<ScalarFeedbackMatrix> mixing_matrices);
 
     void Process(const AudioBuffer& input, AudioBuffer& output) override;
 
-    size_t InputChannelCount() const override
+    uint32_t InputChannelCount() const override
     {
         return N_;
     }
 
-    size_t OutputChannelCount() const override
+    uint32_t OutputChannelCount() const override
     {
         return N_;
     }
