+++ src/schroeder_allpass.h
@@ -13,9 +13,9 @@ namespace sfFDN
 class SchroederAllpass
 {
   public:
-    SchroederAllpass(size_t delay, float g);
+    SchroederAllpass(uint32_t delay, float g);
 
-    void SetDelay(size_t delay);
+    void SetDelay(uint32_t delay);
     void SetG(float g);
 
     float Tick(float input);
@@ -29,19 +29,19 @@ class SchroederAllpass
 class SchroederAllpassSection : public AudioProcessor
 {
   public:
-    SchroederAllpassSection(size_t N);
+    SchroederAllpassSection(uint32_t N);
 
-    void SetDelays(std::span<size_t> delays);
+    void SetDelays(std::span<uint32_t> delays);
     void SetGains(std::span<float> gains);
 
     void Process(const AudioBuffer& input, AudioBuffer& output) override;
 
-    size_t InputChannelCount() const override;
+    uint32_t InputChannelCount() const override;
 
-    size_t OutputChannelCount() const override;
+    uint32_t OutputChannelCount() const override;
 
   private:
     std::vector<SchroederAllpass> allpasses_;
-    size_t stage_;
+    uint32_t stage_;
 };
 } // namespace sfFDN
