+++ tests/delay_tests.cpp
@@ -17,15 +17,15 @@ TEST_CASE("Delay")
     sfFDN::Delay delay(1, 10);
 
     std::vector<float> output;
-    constexpr size_t iteration = 10;
-    for (size_t i = 0; i < iteration; ++i)
+    constexpr uint32_t iteration = 10;
+    for (uint32_t i = 0; i < iteration; ++i)
     {
         output.push_back(delay.Tick(i));
     }
 
     constexpr float expected_output[] = {0, 0, 1, 2, 3, 4, 5, 6, 7, 8};
 
-    for (size_t i = 0; i < iteration; ++i)
+    for (uint32_t i = 0; i < iteration; ++i)
     {
         CHECK(output[i] == doctest::Approx(expected_output[i]).epsilon(0.01));
     }
@@ -36,27 +36,27 @@ TEST_CASE("DelayA")
     sfFDN::DelayAllpass delay(1.5, 10);
 
     std::vector<float> output;
-    constexpr size_t iteration = 10;
-    for (size_t i = 0; i < iteration; ++i)
+    constexpr uint32_t iteration = 10;
+    for (uint32_t i = 0; i < iteration; ++i)
     {
         output.push_back(delay.Tick(i));
     }
 
     constexpr float expected_output[] = {0, 0, 0.33, 1.55, 2.48, 3.50, 4.49, 5.50, 6.50, 7.50};
 
-    for (size_t i = 0; i < iteration; ++i)
+    for (uint32_t i = 0; i < iteration; ++i)
     {
         CHECK(output[i] == doctest::Approx(expected_output[i]).epsilon(0.01));
     }
 }
 
 template <typename T>
-void TestDelayBlock(float delay, size_t block_size, size_t max_delay)
+void TestDelayBlock(float delay, uint32_t block_size, uint32_t max_delay)
 {
     T delay_sample(delay, max_delay);
 
     std::vector<float> output_sample;
-    for (size_t i = 0; i < block_size; ++i)
+    for (uint32_t i = 0; i < block_size; ++i)
     {
         output_sample.push_back(delay_sample.Tick(i));
     }
@@ -72,7 +72,7 @@ void TestDelayBlock(float delay, size_t block_size, size_t max_delay)
 
     delay_block.Process(input_buffer, output_buffer);
 
-    for (size_t i = 0; i < block_size; ++i)
+    for (uint32_t i = 0; i < block_size; ++i)
     {
         CHECK(output_block[i] == doctest::Approx(output_sample[i]));
     }
@@ -90,8 +90,8 @@ TEST_CASE("DelayABlock")
 
 TEST_CASE("DelayBank")
 {
-    constexpr size_t kNumDelay = 4;
-    constexpr std::array<size_t, kNumDelay> delays = {2, 3, 4, 5};
+    constexpr uint32_t kNumDelay = 4;
+    constexpr std::array<uint32_t, kNumDelay> delays = {2, 3, 4, 5};
     sfFDN::DelayBank delay_bank(delays, 10);
 
     std::vector<float> output;
@@ -108,8 +108,8 @@ TEST_CASE("DelayBank")
         output.push_back(i);
     }
 
-    constexpr size_t iter = 9;
-    for (size_t i = 0; i < iter; ++i)
+    constexpr uint32_t iter = 9;
+    for (uint32_t i = 0; i < iter; ++i)
     {
         delay_bank.GetNextOutputs(buffer_audio);
         for (auto& i : buffer)
@@ -127,7 +127,7 @@ TEST_CASE("DelayBank")
     constexpr std::array<float, 10> delay3_expected = {0, 0, 0, 0, 0, 1, 0, 0, 0, 0};
 
     CHECK(output.size() == 40);
-    for (size_t i = 0; i < output.size(); i += 4)
+    for (uint32_t i = 0; i < output.size(); i += 4)
     {
         CHECK(output[i] == doctest::Approx(delay0_expected[i / 4]).epsilon(0.01));
         CHECK(output[i + 1] == doctest::Approx(delay1_expected[i / 4]).epsilon(0.01));
@@ -138,14 +138,14 @@ TEST_CASE("DelayBank")
 
 TEST_CASE("DelayBankBlock")
 {
-    constexpr size_t kNumDelay = 4;
-    constexpr size_t kBlockSize = 8;
-    constexpr std::array<size_t, kNumDelay> delays = {2, 3, 4, 5};
+    constexpr uint32_t kNumDelay = 4;
+    constexpr uint32_t kBlockSize = 8;
+    constexpr std::array<uint32_t, kNumDelay> delays = {2, 3, 4, 5};
     sfFDN::DelayBank delay_bank(delays, 10);
 
     std::vector<float> input(kNumDelay * kBlockSize, 0.f);
     // Input vector is deinterleaved by delay line: {d0_0, d0_1, d0_2, ..., d1_0, d1_1, d1_2, ..., dN_0, dN_1, dN_2}
-    for (size_t i = 0; i < kNumDelay; ++i)
+    for (uint32_t i = 0; i < kNumDelay; ++i)
     {
         input[i * kBlockSize] = 1.f;
     }
@@ -162,19 +162,19 @@ TEST_CASE("DelayBankBlock")
     constexpr std::array<float, kBlockSize> delay2_expected = {0, 0, 0, 0, 1, 0, 0, 0};
     constexpr std::array<float, kBlockSize> delay3_expected = {0, 0, 0, 0, 0, 1, 0, 0};
 
-    for (size_t j = 0; j < kBlockSize; ++j)
+    for (uint32_t j = 0; j < kBlockSize; ++j)
     {
         CHECK(output[0 * kBlockSize + j] == doctest::Approx(delay0_expected[j]).epsilon(0.01));
     }
-    for (size_t j = 0; j < kBlockSize; ++j)
+    for (uint32_t j = 0; j < kBlockSize; ++j)
     {
         CHECK(output[1 * kBlockSize + j] == doctest::Approx(delay1_expected[j]).epsilon(0.01));
     }
-    for (size_t j = 0; j < kBlockSize; ++j)
+    for (uint32_t j = 0; j < kBlockSize; ++j)
     {
         CHECK(output[2 * kBlockSize + j] == doctest::Approx(delay2_expected[j]).epsilon(0.01));
     }
-    for (size_t j = 0; j < kBlockSize; ++j)
+    for (uint32_t j = 0; j < kBlockSize; ++j)
     {
         CHECK(output[3 * kBlockSize + j] == doctest::Approx(delay3_expected[j]).epsilon(0.01));
     }
